# documentation/openapi.json
{
  "components" : {
    "schemas" : {
...
      "RestCorrectionDePresence" : {
        "description" : "La correction d'une presence fausse : une annulation et une regularisation en un seul acte.",
        "properties" : {
          "dateDeSurvenue" : {
            "description" : "Heure metier corrigee.",
            "format" : "date-time",
            "type" : "string"
          },
          "motif" : {
            "description" : "Motif de l'annulation.",
            "example" : "Heure erronee",
            "maxLength" : 255,
            "minLength" : 0,
            "type" : "string"
          },
          "type" : {
            "description" : "Nature du pointage corrige.",
            "enum" : [ "ARRIVEE", "PAUSE", "REPRISE", "DEPART" ],
            "type" : "string"
          }
        },
        "required" : [ "dateDeSurvenue", "motif", "type" ],
        "type" : "object"
      },
...
      "RestJourneeDeTravail" : {
        "description" : "La journee de travail d'un operateur : une venue, bornee par une arrivee et un depart.\n\nCe n'est **pas** un jour calendaire — le contexte ne connait ni fuseau horaire ni date. L'API expose a la fois\n`amplitude` (de l'arrivee au depart) et `fenetres` (pauses retirees) sans choisir laquelle compte pour la paie.\n",
        "properties" : {
          "amplitude" : {
            "$ref" : "#/components/schemas/RestPeriode",
            "description" : "De l'arrivee au depart, pauses comprises. Absente tant que la journee est ouverte."
          },
          "etat" : {
            "description" : "ABSENT, PRESENT ou EN_PAUSE. Deduit du journal.",
            "enum" : [ "ABSENT", "PRESENT", "EN_PAUSE" ],
            "type" : "string"
          },
          "fenetres" : {
            "description" : "Les intervalles de presence effective, pauses retirees.",
            "items" : {
              "$ref" : "#/components/schemas/RestFenetreDePresence"
            },
            "type" : "array"
          },
          "id" : {
            "description" : "Identifiant de la journee.",
            "format" : "uuid",
            "type" : "string"
          },
          "journal" : {
            "description" : "Le journal complet, annules compris, du plus ancien au plus recent.",
            "items" : {
              "$ref" : "#/components/schemas/RestEvenementDePresence"
            },
            "type" : "array"
          },
          "operateur" : {
            "$ref" : "#/components/schemas/RestOperateurDAtelier",
            "description" : "Operateur concerne."
          }
        },
        "type" : "object"
      },
...
      "RestRegularisationDePresence" : {
        "description" : "Une presence rattrapee par le gestionnaire, sur une journee deja ouverte, a l'heure ou elle a reellement eu lieu.\n\nUne seule regularisation de depart suffit a refermer tous les elements restes ouverts ce jour-la : le correctif est\nporte par la presence, jamais recopie element par element.\n",
        "properties" : {
          "dateDeSurvenue" : {
            "description" : "Heure metier a laquelle le fait a reellement eu lieu.",
            "format" : "date-time",
            "type" : "string"
          },
          "type" : {
            "description" : "Nature du pointage rattrape.",
            "enum" : [ "ARRIVEE", "PAUSE", "REPRISE", "DEPART" ],
            "type" : "string"
          }
        },
        "required" : [ "dateDeSurvenue", "type" ],
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearer-jwt" : {
        "bearerFormat" : "JWT",
        "description" : "Keycloak access token carrying the `tenant` claim.",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "API of the GLM manufacturing tracking application.\n\n## Authentication\n\nEvery endpoint expects a Keycloak bearer token (realm `glmproject`, issuer `http://localhost:9080/realms/glmproject`).\n\nThe token **must carry a `tenant` claim** naming a declared company: each company owns a PostgreSQL schema, and the\nwhole `/api/**` surface answers `403` to a token without a known tenant. No endpoint takes the company as a\nparameter — it is always read from the token.\n\n## Roles\n\n- `USER` — shop-floor operator: clocks presence and work, reads.\n- `GESTIONNAIRE` — back-office: everything a `USER` can do, plus committing elements to the shop floor, closing\n  them, and correcting entries (`regularise`, `annule`, `corrige`).\n- `ADMIN` — technical administration (`/api/admin/**`, `/management/**`) only. **It grants no business access.**\n\nDevelopment users (password equal to the login): `gestionnaire.impeccmold`, `user.impeccmold`,\n`gestionnaire.katilys`, `user.katilys`.\n",
    "license" : {
      "name" : "Apache 2.0",
      "url" : "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "title" : "glmproject API",
    "version" : "0.0.1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/atelier/journees/{id}/evenements/{evenementId}" : {
      "put" : {
        "description" : "Une annulation et une regularisation en un seul appel.",
        "operationId" : "corrige_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "evenementId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RestCorrectionDePresence"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestJourneeDeTravail"
                }
              }
            },
            "description" : "Journee ou evenement introuvable."
          },
          "409" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestJourneeDeTravail"
                }
              }
            },
            "description" : "Evenement deja annule, ou transition impossible."
          }
        },
        "summary" : "Corriger une presence fausse",
        "tags" : [ "Atelier - presence des operateurs" ]
      }
    },
...
    "/api/atelier/journees/{id}/regularisations" : {
      "post" : {
        "description" : "Le cas type : l'operateur est rentre chez lui sans pointer son depart, et le gestionnaire le saisit le lendemain a\nl'heure reelle. Cette seule regularisation referme aussi tous les elements restes ouverts.\n",
        "operationId" : "regularise_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RestRegularisationDePresence"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestJourneeDeTravail"
                }
              }
            },
            "description" : "La regularisation est enregistree."
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestJourneeDeTravail"
                }
              }
            },
            "description" : "Journee introuvable."
          },
          "409" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestJourneeDeTravail"
                }
              }
            },
            "description" : "Transition impossible depuis l'etat de presence a cet instant."
          }
        },
        "summary" : "Rattraper une presence oubliee",
        "tags" : [ "Atelier - presence des operateurs" ]
      }
    },
    "/api/atelier/suivis" : {
      "get" : {
        "description" : "Le tableau de l'atelier.\n\nTous les filtres sont facultatifs : l'ecran des operateurs veut tous les elements actifs d'un coup, sans notion de\ndate. Le parametre etats accepte plusieurs valeurs (?etats=EN_ATTENTE&etats=EN_COURS) ; absent, il ne filtre rien.\nLa periode, quand elle est fournie, porte sur la date d'engagement.\n\nChaque ligne conserve l'etat et les activites en cours, mais ne contient pas de journal.\nLe journal complet, annules compris, se consulte via GET /api/atelier/suivis/{id}.\n",
        "operationId" : "list_3",
        "parameters" : [ {
          "in" : "query",
          "name" : "debut",
          "required" : false,
          "schema" : {
            "format" : "date-time",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "fin",
          "required" : false,
          "schema" : {
            "format" : "date-time",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "etats",
          "required" : false,
          "schema" : {
            "items" : {
              "enum" : [ "EN_ATTENTE", "EN_COURS", "INTERROMPU", "CLOTURE" ],
              "type" : "string"
            },
            "type" : "array",
            "uniqueItems" : true
          }
        }, {
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "default" : 0,
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "default" : 20,
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageRestSuiviDAtelierEnGrille"
                }
              }
            },
            "description" : "La page demandee sans les journaux, triee par date d'engagement descendante."
          }
        },
        "summary" : "Lister les elements engages",
        "tags" : [ "Atelier - elements engages" ]
      },
      "post" : {
        "description" : "Geste metier du back-office, distinct de la creation de l'element : tout ce qui est cree n'est pas forcement a\nfaire, et c'est cet acte qui fait apparaitre l'element sur l'ecran des operateurs.\n\nLe nom et le type de l'element sont copies a cet instant ; l'atelier ne les relira plus.\n",
        "operationId" : "engage",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RestEngagement"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestSuiviDAtelier"
                }
              }
            },
            "description" : "L'element est engage."
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestSuiviDAtelier"
                }
              }
            },
            "description" : "Aucun element de fabrication ne porte cet identifiant."
          },
          "409" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestSuiviDAtelier"
                }
              }
            },
            "description" : "Cet element est deja engage et non cloture."
          }
        },
        "summary" : "Engager un element en atelier",
        "tags" : [ "Atelier - elements engages" ]
      }
    },
    "/api/atelier/suivis/{id}" : {
      "get" : {
        "description" : "Le suivi complet, avec son journal, evenements annules compris.",
        "operationId" : "get_4",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestSuiviDAtelier"
                }
              }
            },
            "description" : "Suivi introuvable."
          }
        },
        "summary" : "Consulter un element engage",
        "tags" : [ "Atelier - elements engages" ]
      }
    },
    "/api/atelier/suivis/{id}/cloture" : {
      "delete" : {
        "description" : "Retire la cloture. L'element redevient pointable.",
        "operationId" : "annuleLaCloture",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestSuiviDAtelier"
                }
              }
            },
            "description" : "Suivi introuvable."
          }
        },
        "summary" : "Rouvrir un element cloture",
        "tags" : [ "Atelier - elements engages" ]
      },
      "put" : {
        "description" : "La cloture ne fige rien pour le gestionnaire : regularisation, annulation et correction restent possibles ensuite.\nRappelee sur un element deja cloture, cette route deplace la date de cloture.\n",
        "operationId" : "cloture",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RestCloture"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestSuiviDAtelier"
                }
              }
            },
            "description" : "Suivi introuvable."
          },
          "409" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestSuiviDAtelier"
                }
              }
            },
            "description" : "Le journal porte un evenement posterieur a la date de cloture demandee."
          }
        },
        "summary" : "Cloturer un element, ou deplacer sa cloture",
        "tags" : [ "Atelier - elements engages" ]
      }
    },
    "/api/atelier/suivis/{id}/evenements/{evenementId}" : {
      "put" : {
        "description" : "Troisieme acte : une annulation et une regularisation en un seul appel.",
        "operationId" : "corrige",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "evenementId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RestCorrection"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestSuiviDAtelier"
                }
              }
            },
            "description" : "Suivi, evenement, operateur ou poste de travail introuvable."
          },
          "409" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestSuiviDAtelier"
                }
              }
            },
            "description" : "Operateur non habilite sur ce poste, evenement deja annule, ou transition impossible."
          }
        },
        "summary" : "Corriger une saisie fausse",
        "tags" : [ "Atelier - elements engages" ]
      }
    },
    "/api/atelier/suivis/{id}/evenements/{evenementId}/annulation" : {
      "post" : {
        "description" : "Deuxieme acte de correction. L'evenement reste au journal.",
        "operationId" : "annule",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "evenementId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RestAnnulationDEvenement"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestSuiviDAtelier"
                }
              }
            },
            "description" : "Suivi ou evenement introuvable."
          },
          "409" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestSuiviDAtelier"
                }
              }
            },
            "description" : "Evenement deja annule."
          }
        },
        "summary" : "Annuler une saisie en trop",
        "tags" : [ "Atelier - elements engages" ]
      }
    },
    "/api/atelier/suivis/{id}/pointages" : {
      "post" : {
        "description" : "Le geste de l'operateur, date a l'instant present.\n\nNe jamais pointer ici une pause ou un depart : ce sont des faits de la journee de travail de l'operateur, ecrits\nune seule fois via POST /api/atelier/journees/pointages.\n",
        "operationId" : "pointe",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RestPointage"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestSuiviDAtelier"
                }
              }
            },
            "description" : "Le geste identique est rejoue."
          },
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestSuiviDAtelier"
                }
              }
            },
            "description" : "Le pointage est enregistre."
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestSuiviDAtelier"
                }
              }
            },
            "description" : "Le corps est invalide ou la date de survenue est future."
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestSuiviDAtelier"
                }
              }
            },
            "description" : "Suivi, operateur ou poste de travail introuvable."
          },
          "409" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestSuiviDAtelier"
                }
              }
            },
            "description" : "Operateur non habilite sur ce poste, element cloture, transition impossible ou identifiant reutilise."
          }
        },
        "summary" : "Pointer un debut, une non conformite ou une fin",
        "tags" : [ "Atelier - elements engages" ]
      }
    },
    "/api/atelier/suivis/{id}/regularisations" : {
      "post" : {
        "description" : "Premier des trois actes de correction.",
        "operationId" : "regularise",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RestRegularisation"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestSuiviDAtelier"
                }
              }
            },
            "description" : "La regularisation est enregistree."
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestSuiviDAtelier"
                }
              }
            },
            "description" : "Suivi, operateur ou poste de travail introuvable."
          },
          "409" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestSuiviDAtelier"
                }
              }
            },
            "description" : "Operateur non habilite sur ce poste, transition impossible, ou evenement anterieur a l'engagement."
          }
        },
        "summary" : "Rattraper une saisie oubliee",
        "tags" : [ "Atelier - elements engages" ]
      }
    },
    "/api/atelier/suivis/{id}/temps-effectif" : {
      "get" : {
        "description" : "Les intervalles bruts du journal, ramenes aux fenetres de presence des operateurs.\n\nC'est ici que la pause de midi scinde un travail en deux et qu'un depart referme ce que l'operateur a oublie\nd'arreter, sans qu'aucun de ces faits ait eu besoin d'etre recopie dans le journal de l'element. Un intervalle\nsans fin est encore en cours.\n",
        "operationId" : "tempsEffectif",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/RestIntervalleDActivite"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "Suivi introuvable."
          }
        },
        "summary" : "Lire le temps effectivement passe sur un element",
        "tags" : [ "Atelier - elements engages" ]
      }
    },
...
  },
  "security" : [ {
    "bearer-jwt" : [ ]
  } ],
...
}